requirements.txt
discord.py==2.0.1
mysql-connector-python==8.0.31